backend/lambda_layer/requirements.txt
boto3>=1.34.100
botocore>=1.34.100
PyMySQL==1.1.0
cryptography==41.0.7
Pillow==10.0.1